move ShardingSphereRulesBuilder to shardingsphere-common
